// src/app/Interfaces/about-me.ts
export interface IAboutMeOBJ {
  titleEn: string;
  titleAr: string;
  titleFn: string;
  descEn: string;
  descAr: string;
  descFn: string;
  imgPath: string;
  profileEn: string;
  profileFn: string;
  profileAr: string;
  fullNameEn: string;
  fullNameAr: string;
  fullNameFn: string;
  phone: string;
  birthDateEn: string;
  birthDateAr: string;
  birthDateFn: string;
  degreeEn: string;
  degreeFn: string;
  degreeAr: string;
  nationalityEn: string;
  nationalityAr: string;
  nationalityFn: string;
  militaryStatusEn: string;
  militaryStatusAr: string;
  militaryStatusFn: string;
  maritalStatusEn: string;
  maritalStatusAr: string;
  maritalStatusFn: string;
  email: string;
}

export interface IAboutMeView {
  title: string;
  desc: string;
  imgPath: string;
  profile: string;
  fullName: string;
  phone: string;
  birthDate: string;
  degree: string;
  nationality: string;
  militaryStatus: string;
  maritalStatus: string;
  email: string;
}